Question2/requirements.txt
streamlit==1.35.0
numpy==1.26.4
scikit-learn==1.4.2
google-generativeai==0.4.1
requests==2.31.0